fix: add styles, & snapshot for ManageNetworks Component
