+++ app/components/UI/ManageNetworks/ManageNetworks.test.js
@@ -1,0 +1,41 @@
+// Third party dependencies.
+import React from 'react';
+
+// Internal dependencies.
+import ManageNetworks from './ManageNetworks';
+import renderWithProvider from '../../../util/test/renderWithProvider';
+import { useNavigation } from '@react-navigation/native';
+import { useSelector } from 'react-redux';
+import { selectProviderConfig } from '../../../selectors/networkController';
+
+jest.mock('@react-navigation/native', () => {
+  const actualReactNavigation = jest.requireActual('@react-navigation/native');
+  return {
+    ...actualReactNavigation,
+    useNavigation: () => ({
+      navigate: jest.fn(),
+    }),
+  };
+});
+
+jest.mock('react-redux', () => ({
+  ...jest.requireActual('react-redux'),
+  useSelector: jest.fn(),
+}));
+
+const mockProviderConfig = {
+  type: 'mainnet',
+  chainId: '1',
+};
+
+describe('ManageNetworks', () => {
+  it('should render correctly', () => {
+    useSelector.mockImplementation((selector) => {
+      if (selector === selectProviderConfig) return mockProviderConfig;
+    });
+    const { toJSON } = renderWithProvider(
+      <ManageNetworks navigation={useNavigation()} />,
+    );
+    expect(toJSON()).toMatchSnapshot();
+  });
+});
