+++ app/components/UI/ManageNetworks/ManageNetworks.tsx
@@ -1,5 +1,5 @@
 import React, { useCallback } from 'react';
-import { View, StyleSheet, Linking } from 'react-native';
+import { View, Linking } from 'react-native';
 import Text, {
   TextVariant,
   TextColor,
@@ -20,28 +20,7 @@ import { useMetrics } from '../../../components/hooks/useMetrics';
 import { MetaMetricsEvents } from '../../../core/Analytics';
 import { ConnectedAccountsSelectorsIDs } from '../../../../e2e/selectors/Modals/ConnectedAccountModal.selectors';
 import AppConstants from '../../../core/AppConstants';
-
-const styles = StyleSheet.create({
-  setting: {
-    marginVertical: 16,
-  },
-  heading: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-  },
-  description: {
-    fontSize: 14,
-    textAlign: 'left',
-    marginTop: 10,
-    lineHeight: 22,
-    fontWeight: '400',
-  },
-  networkPicker: {
-    marginVertical: 16,
-    alignSelf: 'flex-start',
-  },
-});
+import styles from './ManageNetworks.styles';
 
 export default function ManageNetworksComponent() {
   const providerConfig: ProviderConfig = useSelector(selectProviderConfig);
